klient/terminal: add stub for Windows
