+++ go/src/koding/klient/terminal/pty/pty_windows.go
@@ -1,0 +1,1 @@
+package pty
