+++ go/src/koding/klient/terminal/server.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package terminal
 
 import (
