+++ go/src/koding/klient/terminal/terminal_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package terminal
 
 import (
