+++ go/src/koding/klient/terminal/user.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package terminal
 
 import (
